third_party/blink/web_tests/fast/dom/HTMLImageElement/image-picture-1x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("canary").clientWidth is 75
PASS currentSrcFileName("canary") is "blue_rect.jpg?0"
PASS document.getElementById("simple").clientWidth is 800
PASS currentSrcFileName("simple") is "image-set-2x.png?2"
PASS document.getElementById("multiple").clientWidth is 800
PASS currentSrcFileName("multiple") is "image-set-2x.png?2"
PASS document.getElementById("other_children").clientWidth is 800
PASS currentSrcFileName("other_children") is "image-set-2x.png?2"
PASS document.getElementById("img_before_source").clientWidth is 75
PASS currentSrcFileName("img_before_source") is "blue_rect.jpg?1"
PASS document.getElementById("unknown_type").clientWidth is 1600
PASS currentSrcFileName("unknown_type") is "image-set-4x.png?3"
PASS document.getElementById("svg_type").clientWidth is 800
PASS currentSrcFileName("svg_type") is "image-set-2x.png?2"
PASS document.getElementById("empty_type").clientWidth is 800
PASS currentSrcFileName("empty_type") is "image-set-2x.png?2"
PASS document.getElementById("whitespace_type").clientWidth is 800
PASS currentSrcFileName("whitespace_type") is "image-set-2x.png?2"
PASS document.getElementById("parameter_type").clientWidth is 800
PASS currentSrcFileName("parameter_type") is "image-set-2x.png?2"
PASS document.getElementById("empty_srcset").clientWidth is 1600
PASS currentSrcFileName("empty_srcset") is "image-set-4x.png?3"
PASS document.getElementById("no_srcset").clientWidth is 1600
PASS currentSrcFileName("no_srcset") is "image-set-4x.png?3"
   
some text
        
